blog/requirements.txt
Django==3.2.15
pytz==2022.2.1
markdown2==2.4.3
djangorestframework==3.13.1
